requirements.txt
flask
requests
azure.data.tables
pytz